Make tests work under Python 2. Move fake BZ2 class from test_*.py files into mocking and make it able to act as a context manager.
