+++ test/blast/test_alignments.py
@@ -16,7 +16,7 @@ except ImportError:
 from six import StringIO
 from Bio import SeqIO
 
-from ..mocking import mockOpen
+from ..mocking import mockOpen, File
 from .sample_data import PARAMS, RECORD0, RECORD1, RECORD2, RECORD3, RECORD4
 
 from dark.reads import Read, Reads
@@ -26,27 +26,6 @@ from dark.blast.alignments import (
     BlastReadsAlignments,  ZERO_EVALUE_UPPER_RANDOM_INCREMENT)
 from dark.titles import TitlesAlignments
 from dark import ncbidb
-
-
-class BZ2(object):
-    """
-    A BZ2File mock.
-    """
-    def __init__(self, data):
-        self._data = data
-        self._index = 0
-
-    def close(self):
-        pass
-
-    def readline(self):
-        self._index += 1
-        return self._data[self._index - 1]
-
-    def __iter__(self):
-        index = self._index
-        self._index = len(self._data)
-        return iter(self._data[index:])
 
 
 class TestBlastReadsAlignments(TestCase):
@@ -235,7 +214,7 @@ class TestBlastReadsAlignments(TestCase):
         If a compressed (bz2) JSON file contains a parameters section and one
         record, it must be read correctly.
         """
-        result = BZ2([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+        result = File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
 
         with patch.object(bz2, 'BZ2File') as mockMethod:
             mockMethod.return_value = result
@@ -259,9 +238,9 @@ class TestBlastReadsAlignments(TestCase):
             def sideEffect(self, _ignoredFilename):
                 if self.first:
                     self.first = False
-                    return BZ2([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
                 else:
-                    return BZ2([dumps(PARAMS) + '\n', dumps(RECORD1) + '\n'])
+                    return File([dumps(PARAMS) + '\n', dumps(RECORD1) + '\n'])
 
         sideEffect = SideEffect()
         with patch.object(bz2, 'BZ2File') as mockMethod:
@@ -293,14 +272,14 @@ class TestBlastReadsAlignments(TestCase):
                 if self.count == 0:
                     self.test.assertEqual('1.json.bz2', filename)
                     self.count += 1
-                    return BZ2([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
                 elif self.count == 1:
                     self.test.assertEqual('2.json.bz2', filename)
                     self.count += 1
-                    return BZ2([dumps(PARAMS) + '\n', dumps(RECORD1) + '\n'])
+                    return File([dumps(PARAMS) + '\n', dumps(RECORD1) + '\n'])
                 else:
                     self.test.assertEqual('3.json.bz2', filename)
-                    return BZ2([dumps(PARAMS) + '\n', dumps(RECORD2) + '\n'])
+                    return File([dumps(PARAMS) + '\n', dumps(RECORD2) + '\n'])
 
         sideEffect = SideEffect(self)
         with patch.object(bz2, 'BZ2File') as mockMethod:
@@ -335,11 +314,11 @@ class TestBlastReadsAlignments(TestCase):
             def sideEffect(self, _ignoredFilename):
                 if self.first:
                     self.first = False
-                    return BZ2([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
                 else:
                     params = deepcopy(PARAMS)
                     params['application'] = 'Skype'
-                    return BZ2([dumps(params) + '\n', dumps(RECORD1) + '\n'])
+                    return File([dumps(params) + '\n', dumps(RECORD1) + '\n'])
 
         sideEffect = SideEffect()
         with patch.object(bz2, 'BZ2File') as mockMethod:
@@ -410,7 +389,7 @@ class TestBlastReadsAlignments(TestCase):
         The adjustHspsForPlotting function must alter HSPs so that non-zero
         evalues are converted to the positive value of their negative exponent.
         """
-        result = lambda a: BZ2([
+        result = lambda a: File([
             dumps(PARAMS) + '\n', dumps(deepcopy(RECORD0)) + '\n',
             dumps(deepcopy(RECORD1)) + '\n', dumps(deepcopy(RECORD2)) + '\n',
             dumps(deepcopy(RECORD3)) + '\n'])
@@ -436,7 +415,7 @@ class TestBlastReadsAlignments(TestCase):
         The adjustHspsForPlotting function must alter HSPs so that zero
         evalues are set randomly high.
         """
-        result = lambda a: BZ2([
+        result = lambda a: File([
             dumps(PARAMS) + '\n', dumps(deepcopy(RECORD0)) + '\n',
             dumps(deepcopy(RECORD1)) + '\n', dumps(deepcopy(RECORD2)) + '\n',
             dumps(deepcopy(RECORD3)) + '\n', dumps(deepcopy(RECORD4)) + '\n'])
@@ -1262,7 +1241,7 @@ class TestBlastReadsAlignmentsFiltering(TestCase):
         """
         It must be possible to clear any filtering that has been applied.
         """
-        result = lambda a: BZ2([
+        result = lambda a: File([
             dumps(PARAMS) + '\n', dumps(RECORD0) + '\n',
             dumps(RECORD1) + '\n', dumps(RECORD2) + '\n'])
